Add initial EdDSA verification support.
This adds support for verifying EdDSA signatures, when using OpenSSL.
Due to the way EdDSA works, an auxiliary buffer is need to serialize
the Sig_Structure. A new function, t_cose_sign1_verify_set_sigstruct_buffer,
is used by the caller to provide this buffer.

A "known good" signature, generated by pycose, is added to the test
suite.